feat: inserter for work-order
