+++ .golangci.yaml
@@ -60,7 +60,7 @@ linters-settings:
         - github.com/prometheus/client_golang/prometheus/promauto
         - github.com/prometheus/client_golang/prometheus/promhttp
         - github.com/rung/go-safecast
-        - github.com/united-manufacturing-hub/*/*
+        - github.com/united-manufacturing-hub/**/*
         - github.com/zeebo/xxh3
         - go.uber.org/zap
         - golang.org/*/*
