+++ golang/cmd/kafka-to-postgresql-v2/postgresql/lru.go
@@ -1,0 +1,194 @@
+package postgresql
+
+import (
+	"encoding/binary"
+	"errors"
+	"github.com/jackc/pgx/v5"
+	sharedStructs "github.com/united-manufacturing-hub/united-manufacturing-hub/cmd/kafka-to-postgresql-v2/shared"
+	"golang.org/x/crypto/sha3"
+	"strings"
+	"sync"
+)
+
+var goiLock = sync.Mutex{}
+
+func (c *Connection) GetOrInsertAsset(topic *sharedStructs.TopicDetails) (uint64, error) {
+	if c.db == nil {
+		return 0, errors.New("database is nil")
+	}
+	id, hit := c.lookupAssetIdLRU(topic)
+	if hit {
+		return id, nil
+	}
+
+	goiLock.Lock()
+	defer goiLock.Unlock()
+	// It might be that another locker already added it, so we do a double check
+	id, hit = c.lookupAssetIdLRU(topic)
+	if hit {
+		return id, nil
+	}
+
+	selectQuery := `SELECT id FROM asset WHERE enterprise = $1 AND 
+    site = $2 AND 
+    area = $3 AND 
+    line = $4 AND 
+    workcell = $5 AND 
+    origin_id = $6`
+	selectRowContext, selectRowContextCncl := get1MinuteContext()
+	defer selectRowContextCncl()
+
+	var err error
+	err = c.db.QueryRow(selectRowContext, selectQuery, topic.Enterprise, topic.Site, topic.Area, topic.ProductionLine, topic.WorkCell, topic.OriginId).Scan(&id)
+	if err != nil {
+		if errors.Is(err, pgx.ErrNoRows) {
+			// Row isn't found, need to insert
+			insertQuery := `INSERT INTO asset (enterprise, site, area, line, workcell, origin_id) VALUES ($1, $2, $3, $4, $5, $6) RETURNING id`
+			insertRowContext, insertRowContextCncl := get1MinuteContext()
+			defer insertRowContextCncl()
+
+			err = c.db.QueryRow(insertRowContext, insertQuery, topic.Enterprise, topic.Site, topic.Area, topic.ProductionLine, topic.WorkCell, topic.OriginId).Scan(&id)
+			if err != nil {
+				return 0, err
+			}
+
+			// Add to LRU cache
+			c.addToAssetIdLRU(topic, id)
+
+			return id, nil
+		} else {
+			return 0, err
+		}
+	}
+
+	// If no error and asset exists
+	c.addToAssetIdLRU(topic, id)
+	return id, nil
+}
+
+func (c *Connection) addToAssetIdLRU(topic *sharedStructs.TopicDetails, id uint64) {
+	c.assetIdCache.Add(getCacheKeyFromTopic(topic), id)
+}
+
+func (c *Connection) lookupAssetIdLRU(topic *sharedStructs.TopicDetails) (uint64, bool) {
+	value, ok := c.assetIdCache.Get(getCacheKeyFromTopic(topic))
+	if ok {
+		c.lruHits.Add(1)
+		return value.(uint64), true
+	}
+	c.lruMisses.Add(1)
+	return 0, false
+}
+
+var ptIdLock = sync.Mutex{}
+
+func (c *Connection) GetOrInsertProductType(assetId uint64, product sharedStructs.WorkOrderCreateMessageProduct) (uint64, error) {
+	/*
+		CREATE TABLE product_types (
+		    productTypeId INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+		    externalProductTypeId TEXT NOT NULL,
+		    cycleTime INTEGER NOT NULL,
+		    assetId INTEGER REFERENCES assets(id),
+		    CONSTRAINT external_product_asset_uniq UNIQUE (externalProductTypeId, assetId),
+		    CHECK (cycleTime > 0)
+		);
+	*/
+	if c.db == nil {
+		return 0, errors.New("database is nil")
+	}
+	ptId, hit := c.lookupProductTypeIdLRU(assetId, product)
+	if hit {
+		return ptId, nil
+	}
+
+	ptIdLock.Lock()
+	defer ptIdLock.Unlock()
+
+	// It might be that another locker already added it, so we do a double check
+	ptId, hit = c.lookupProductTypeIdLRU(assetId, product)
+	if hit {
+		return ptId, nil
+	}
+
+	selectQuery := `SELECT productTypeId FROM product_types WHERE externalProductTypeId = $1 AND assetId = $2`
+	selectRowContext, selectRowContextCncl := get1MinuteContext()
+	defer selectRowContextCncl()
+
+	var err error
+	err = c.db.QueryRow(selectRowContext, selectQuery, product.ExternalProductId, assetId).Scan(&ptId)
+	if err != nil {
+		if errors.Is(err, pgx.ErrNoRows) {
+			// Row isn't found, need to insert
+			insertQuery := `INSERT INTO product_types (externalProductTypeId, cycleTime, assetId) VALUES ($1, $2, $3) RETURNING productTypeId`
+			insertRowContext, insertRowContextCncl := get1MinuteContext()
+			defer insertRowContextCncl()
+
+			err = c.db.QueryRow(insertRowContext, insertQuery, product.ExternalProductId, product.CycleTimeMs, assetId).Scan(&ptId)
+			if err != nil {
+				return 0, err
+			}
+
+			// Add to LRU cache
+			c.addToProductTypeIdLRU(assetId, product, ptId)
+
+			return ptId, nil
+		} else {
+			return 0, err
+		}
+	}
+
+	// If no error and product type exists
+	c.addToProductTypeIdLRU(assetId, product, ptId)
+	return ptId, nil
+}
+
+func (c *Connection) addToProductTypeIdLRU(assetId uint64, product sharedStructs.WorkOrderCreateMessageProduct, ptId uint64) {
+	c.productTypeIdCache.Add(getCacheKeyFromProduct(assetId, product), ptId)
+}
+
+func (c *Connection) lookupProductTypeIdLRU(assetId uint64, product sharedStructs.WorkOrderCreateMessageProduct) (uint64, bool) {
+	value, ok := c.productTypeIdCache.Get(getCacheKeyFromProduct(assetId, product))
+	if ok {
+		c.lruHits.Add(1)
+		return value.(uint64), true
+	}
+	c.lruMisses.Add(1)
+	return 0, false
+}
+
+func getCacheKeyFromTopic(topic *sharedStructs.TopicDetails) string {
+	// Attempt cache lookup
+	var cacheKey strings.Builder
+	cacheKey.WriteString(topic.Enterprise)
+	cacheKey.WriteRune('*') // This char cannot occur in the topic, and therefore can be safely used as a seperator
+	cacheKey.WriteRune('s')
+	cacheKey.WriteString(topic.Site)
+
+	cacheKey.WriteRune('*') // This char cannot occur in the topic, and therefore can be safely used as a seperator
+	cacheKey.WriteRune('a')
+	cacheKey.WriteString(topic.Area)
+
+	cacheKey.WriteRune('*') // This char cannot occur in the topic, and therefore can be safely used as a seperator
+	cacheKey.WriteRune('p')
+	cacheKey.WriteString(topic.ProductionLine)
+
+	cacheKey.WriteRune('*') // This char cannot occur in the topic, and therefore can be safely used as a seperator
+	cacheKey.WriteRune('w')
+	cacheKey.WriteString(topic.WorkCell)
+
+	cacheKey.WriteRune('*') // This char cannot occur in the topic, and therefore can be safely used as a seperator
+	cacheKey.WriteRune('o')
+	cacheKey.WriteString(topic.OriginId)
+	return cacheKey.String()
+}
+
+func getCacheKeyFromProduct(assetId uint64, product sharedStructs.WorkOrderCreateMessageProduct) string {
+	// Attempt cache lookup
+	hasher := sha3.New512()
+	var assetIdBytes []byte
+	binary.LittleEndian.PutUint64(assetIdBytes, assetId)
+	hasher.Write(assetIdBytes)
+	hasher.Write([]byte(product.ExternalProductId))
+
+	return string(hasher.Sum(nil))
+}
